refactor(emitter): split top-level using class assignment helpers
